Massive refactoring Deckerization
