+++ src/main/java/am/aua/cs322/orgchart/AuthServiceApplication.java
@@ -1,0 +1,13 @@
+package am.aua.cs322.orgchart;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AuthServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(AuthServiceApplication.class, args);
+	}
+
+}
