+++ src/main/java/am/aua/cs322/orgchart/errors/ErrorMessages.java
@@ -1,0 +1,5 @@
+package am.aua.cs322.orgchart.errors;
+
+public enum ErrorMessages {
+    USER_DISABLED, INVALID_CREDENTIALS, MISSING_DATA
+}
