+++ src/main/java/am/aua/cs322/orgchart/exceptions/AuthenticationException.java
@@ -1,0 +1,7 @@
+package am.aua.cs322.orgchart.exceptions;
+
+public class AuthenticationException extends RuntimeException {
+    public AuthenticationException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
