+++ src/main/java/am/aua/cs322/orgchart/filters/JwtTokenAuthorizationOncePerRequestFilter.java
@@ -1,0 +1,84 @@
+package am.aua.cs322.orgchart.filters;
+
+import am.aua.cs322.orgchart.utils.JwtTokenUtil;
+import io.jsonwebtoken.ExpiredJwtException;
+import org.slf4j.Logger;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
+import org.springframework.stereotype.Component;
+import org.springframework.util.AntPathMatcher;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+import javax.servlet.FilterChain;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.util.List;
+
+import static org.slf4j.LoggerFactory.getLogger;
+
+@Component
+public class JwtTokenAuthorizationOncePerRequestFilter extends OncePerRequestFilter {
+    private final Logger logger = getLogger(this.getClass());
+
+    @Autowired
+    @Qualifier("inMemoryUserDetailsService")
+    private UserDetailsService jwtInMemoryUserDetailsService;
+
+    @Autowired
+    private JwtTokenUtil jwtTokenUtil;
+
+    @Value("${jwt.http.request.header}")
+    private String tokenHeader;
+
+    private final AntPathMatcher matcher = new AntPathMatcher();
+    private final List<String> whiteList = List.of("/authenticate");
+
+    @Override
+    protected boolean shouldNotFilter(HttpServletRequest request) throws ServletException {
+        return whiteList.stream().anyMatch(s -> matcher.match(s, request.getServletPath()));
+    }
+
+    @Override
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain chain) throws ServletException, IOException {
+        logger.debug("Authentication Request For '{}'", request.getRequestURL());
+
+        final String requestTokenHeader = request.getHeader(this.tokenHeader);
+
+        String username = null;
+        String jwtToken = null;
+        if (requestTokenHeader != null && requestTokenHeader.startsWith("Bearer ")) {
+            jwtToken = requestTokenHeader.substring(7);
+            try {
+                username = jwtTokenUtil.getUsernameFromToken(jwtToken);
+            } catch (IllegalArgumentException e) {
+                logger.error("JWT_TOKEN_UNABLE_TO_GET_USERNAME", e);
+            } catch (ExpiredJwtException e) {
+                logger.warn("JWT_TOKEN_EXPIRED", e);
+            }
+        } else {
+            logger.warn("JWT_TOKEN_DOES_NOT_START_WITH_BEARER_STRING");
+        }
+
+        logger.debug("JWT_TOKEN_USERNAME_VALUE '{}'", username);
+        if (username != null && SecurityContextHolder.getContext().getAuthentication() == null) {
+
+            UserDetails userDetails = this.jwtInMemoryUserDetailsService.loadUserByUsername(username);
+
+            if (jwtTokenUtil.validateToken(jwtToken, userDetails)) {
+                UsernamePasswordAuthenticationToken usernamePasswordAuthenticationToken = new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
+                usernamePasswordAuthenticationToken.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
+                SecurityContextHolder.getContext().setAuthentication(usernamePasswordAuthenticationToken);
+            }
+        }
+
+        chain.doFilter(request, response);
+    }
+}
