+++ src/main/java/am/aua/cs322/orgchart/models/User.java
@@ -1,0 +1,77 @@
+package am.aua.cs322.orgchart.models;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import lombok.*;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import java.util.Collection;
+import java.util.UUID;
+
+/**
+ * User POJO
+ */
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@EqualsAndHashCode
+public class User implements UserDetails {
+    @JsonIgnore
+    private UUID id;
+    private String email;
+    private String username;
+
+    @JsonIgnore
+    private String password;
+    private String position;
+    private String department;
+    private String location;
+    private String firstName;
+    private String lastName;
+    private String phone;
+    private String photoUrl;
+
+
+    public String getUsername() {
+        return username;
+    }
+
+    @JsonIgnore
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @JsonIgnore
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @JsonIgnore
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @JsonIgnore
+    @Override
+    public boolean isEnabled() {
+        return true;
+    }
+
+
+    @JsonIgnore
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return null;
+    }
+
+    @JsonIgnore
+    @Override
+    public String getPassword() {
+        return password;
+    }
+}
