--- src/main/java/com/cs322/errors/ErrorMessages.java
@@ -1,5 +1,0 @@
-package com.cs322.errors;
-
-public enum ErrorMessages {
-    USER_DISABLED, INVALID_CREDENTIALS, MISSING_DATA
-}
