--- src/main/java/com/cs322/exceptions/AuthenticationException.java
@@ -1,7 +1,0 @@
-package com.cs322.exceptions;
-
-public class AuthenticationException extends RuntimeException {
-    public AuthenticationException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
