--- src/main/java/com/cs322/models/JsonError.java
@@ -1,13 +1,0 @@
-package com.cs322.models;
-
-import lombok.*;
-
-import java.io.Serializable;
-
-@Getter
-@Setter
-@RequiredArgsConstructor
-public class JsonError implements Serializable {
-    @NonNull
-    private String message;
-}
