--- src/main/java/com/cs322/models/JwtTokenResponse.java
@@ -1,29 +1,0 @@
-package com.cs322.models;
-
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.io.Serializable;
-
-/**
- * JWT Token Response POJO
- */
-
-@Getter
-@Setter
-@EqualsAndHashCode
-public class JwtTokenResponse implements Serializable {
-
-    private static final long serialVersionUID = 8317676219297719109L;
-
-    private final String token;
-
-    public JwtTokenResponse(String token) {
-        this.token = token;
-    }
-
-    public String getToken() {
-        return this.token;
-    }
-}
